Weekly Development Merge v0.0.2

+++ src/game.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <sstream>
 #include "game.h"
-#include "utils.h"
+#include "console/utils.h"
 
 using namespace std;
 
@@ -77,7 +77,7 @@ void Game::update() {
 	draw();
 	string in;
 	cin >> in;
-	unsigned int choice = 0;
+	size_t choice = 0;
 	istringstream(in) >> choice;
     if(player.get_mana() > 0 && choice != 0) {
         player.cast(choice - 1);
